+++ alacritty/src/window_context.rs
@@ -325,6 +325,9 @@ impl WindowContext {
         }
 
         if self.dirty {
+            // Force the display to process any pending display update.
+            self.display.process_renderer_update();
+
             self.dirty = false;
 
             // Request immediate re-draw if visual bell animation is not finished yet.